% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.osm.food-package.R
\docType{package}
\name{dc.osm.food-package}
\alias{dc.osm.food}
\alias{dc.osm.food-package}
\title{dc.osm.food: Social Data Commons: Health, Behavior, Diet: Modified Retail Food Environment Index}
\description{
Allows user to easily get retail food environment index data from the SDAD Data Commons.  Data can be provided directly or in a file.
}
\author{
\strong{Maintainer}: Joanna Schroeder \email{js2mr@virginia.edu}

Authors:
\itemize{
  \item Social and Decision Analytics Division (SDAD), Biocomplexity Institute, University of Virginia
}

}
\keyword{internal}
